+++ smartcontract/sdk/go/telemetry/constants.go
@@ -12,6 +12,8 @@ const (
 	InitializeInternetLatencySamplesInstructionIndex TelemetryInstructionType = 2
 	// Represents the write internet latency samples instruction
 	WriteInternetLatencySamplesInstructionIndex TelemetryInstructionType = 3
+	// Represents the initialize timestamp index instruction
+	InitializeTimestampIndexInstructionIndex TelemetryInstructionType = 4
 
 	// InstructionErrorAccountSamplesAccountFull is the error code that the telemetry program returns
 	// when the given PDA has reached maximum capacity for samples.
@@ -20,6 +22,10 @@ const (
 	// InstructionErrorAccountDoesNotExist is the error code that the telemetry program returns
 	// when the given PDA does not exist.
 	InstructionErrorAccountDoesNotExist = 1011
+
+	// InstructionErrorTimestampIndexAccountDoesNotExist is the error code that the telemetry
+	// program returns when the timestamp index account does not exist.
+	InstructionErrorTimestampIndexAccountDoesNotExist = 1019
 
 	// MaxSamplesPerBatch is the maximum number of samples that can be written in a single batch.
 	//
@@ -58,4 +64,6 @@ const (
 	DeviceLatencySamplesSeed = "dzlatency"
 	// Seed for internet latency samples PDAs
 	InternetLatencySamplesSeed = "inetlatency"
+	// Seed for timestamp index PDAs
+	TimestampIndexSeed = "tsindex"
 )